==> notebooks/01_split.ipynb <==
import pandas as pd
import numpy as np
from sklearn.model_selection import train_test_split

# load the data
bank_data = pd.read_csv('../data/bank-additional-full.csv', sep=";")

bank_data.head()

## remove duplicate rows

bank_data.shape

bank_data_unique = bank_data.drop_duplicates()

bank_data_unique.shape

## split the data into train, validation, and test sets
using a 70/15/15 split

# first split the test set from the train and validation set
train_valid, test_data = train_test_split(bank_data_unique, test_size = 0.14999)

# then split the train and validation sets
train_data, valid_data = train_test_split(train_valid, test_size = 0.17645)

# save as csv files -- these file paths are for this notebook only

test_data.to_csv('../data/test_data.csv', index = False)
print('shape of unprocessed test data: ', test_data.shape)

valid_data.to_csv('../data/valid_data.csv', index = False)
print('shape of unprocessed validation data: ', valid_data.shape)

train_data.to_csv('../data/train_data.csv', index = False)
print('shape of unprocessed training data: ', train_data.shape)
