==> notebooks/02_process.ipynb <==
import pandas as pd
import numpy as np
from joblib import dump, load
from sklearn.preprocessing import OneHotEncoder

# load the data to be processed
train_data = pd.read_csv('../data/train_data.csv')
valid_data = pd.read_csv('../data/valid_data.csv')
test_data = pd.read_csv('../data/test_data.csv')

## create functions for one-hot encoding categorical columns

def create_ohe(train_data, filename='joblib/ohe.joblib'):
    """
    Return an one-hot encoder fitted using `train_data` as a joblib file.
    Parameters
    ----------
    train_data: pandas.core.frame.DataFrame
        Training set
    filename: str
        Filename to use to save encoder
    """

    # identify columns to one-hot encode
    features = train_data.drop(columns=['y'])
    to_encode = features.columns[features.dtypes==object].tolist()

    # create encoder
    ohe = OneHotEncoder(handle_unknown='ignore', sparse=False, dtype=int)

    # fit the encoder on training data
    ohe.fit(train_data.loc[:, to_encode])

    # check number of columns to encode
    assert len(ohe.categories_) == len(to_encode)

    # Save the OHE
    dump(ohe, filename)


    return None

def apply_ohe(data, filename='ohe.joblib'):
    """
    Given a one-hot encoder fit on `train_data`, return a data frame
    with the one-hot encoder applied.
    Parameters
    ----------
    data: pandas.core.frame.DataFrame
        `train_data`, `valid_data` or `test_data`
    filename: str
        Filename of fitted encoder
    Returns
    -------
    pandas.core.frame.DataFrame
    """

    # WARNING: `to_encode` must match list passed in create_ohe
    features = data.drop(columns=['y'])
    to_encode = features.columns[features.dtypes==object].tolist()

    # load the encoder that has been fit on the training data
    ohe = load(filename)

    # apply the encoder
    ohe_array = ohe.transform(data.loc[:, to_encode])

    # get names of encoded columns and add variable name as prefix
    ohe_cols = []
    for i in range(len(ohe.categories_)):
        ohe_cols.append(to_encode[i] + '_' + ohe.categories_[i])

    ohe_cols = np.concatenate(ohe_cols).ravel()

    # create dataframe containing encoded columns (preserve original row indices)
    ohe_df = pd.DataFrame(ohe_array, index=data.index, columns=ohe_cols)

    # concatenate with existing data frame
    data_ohe = pd.concat((data, ohe_df), axis=1)

    # drop the columns that were one-hot encoded
    data_ohe = data_ohe.drop(columns = to_encode)

    # check that the number of rows is unchanged
    assert data_ohe.shape[0] == data.shape[0]

    # check that first column in `to_encode` is not in the encoded dataframe
    assert to_encode[0] not in data_ohe.columns.to_list()
    
    # drop `y` from encoded data
    data_ohe = data_ohe.drop(columns = ['y'])


    return data_ohe

# apply_ohe(train_data, filename='joblib/ohe.joblib').info()

## pre-process the training data

# create a function to pre-process the data

def process_train(train_data):
    """
    Processes training data set. 
    Fits and applies a one-hot encoder. 
    Drops some features. 
    Makes target variable numeric & binary.
    Returns a dataframe.
    
    Parameters
    ----------
    train_data : pandas.core.frame.DataFrame
       The training data set ('train_data.csv')

    Returns
    -------
    pandas.core.frame.DataFrame
        The processed train set with features + target 
    """
        
    # target
    y = train_data['y'].replace({'no':0, 'yes':1})
    
    # predictors
    features = train_data.drop(columns=['y'])
    
    # fit and save one-hot encoder
    create_ohe(train_data, filename='joblib/ohe.joblib')
    
    # transform data using one-hot encoder
    data_ohe = apply_ohe(train_data, filename='joblib/ohe.joblib')
    
    # drop `duration`
    data_ohe = data_ohe.drop(columns = ['duration'])
    
#     # drop highly correlated features
#     data_ohe = data_ohe.drop(columns = ['emp.var.rate', 'euribor3m'])
    
    # combine processed predictor data back with target
    data = pd.concat([y, data_ohe], axis=1, sort='false')
    
    
    return data

train_data_processed = process_train(train_data)
train_data_processed

train_data_processed.info()

## pre-process the validation and test data

# create a function to pre-process the data

def process_valid_test(valid_test_data):
    """
    Processes validation and/or test data sets. 
    Applies a one-hot encoder that was fit on training data.
    Drops some features. 
    Makes target variable numeric & binary.
    Returns a dataframe.
    
    Parameters
    ----------
    valid_test_data : pandas.core.frame.DataFrame
       Either the validation or the test data set ('valid_data.csv' or 'test_data.csv')

    Returns
    -------
    pandas.core.frame.DataFrame
        Either the processed validation or test set with features + target 
    """
        
    # target
    y = valid_test_data['y'].replace({'no':0, 'yes':1})
    
    # predictors
    
    # transform data using saved one-hot encoder
    data_ohe = apply_ohe(valid_test_data, filename='joblib/ohe.joblib')
    
    # drop `duration`
    data_ohe = data_ohe.drop(columns = ['duration'])
    
#     # drop highly correlated features
#     data_ohe = data_ohe.drop(columns = ['emp.var.rate', 'euribor3m'])
    
    # combine processed predictor data back with target
    data = pd.concat([y, data_ohe], axis=1, sort='false')
    
    
    return data

valid_data_processed = process_valid_test(valid_data)
valid_data_processed

test_data_processed = process_valid_test(test_data)
test_data_processed

## save as csv files

# save as csv files
train_data_processed.to_csv('../data/train_data_processed.csv', index = False)
print('shape of processed train data: ', train_data_processed.shape)

valid_data_processed.to_csv('../data/valid_data_processed.csv', index = False)
print('shape of processed validation data: ', valid_data_processed.shape)

test_data_processed.to_csv('../data/test_data_processed.csv', index = False)
print('shape of processed test data: ', test_data_processed.shape)
