==> notebooks/03_model.ipynb <==
Test several models using default parameters.

import pandas as pd
import numpy as np
import matplotlib.pyplot as plt 

from sklearn.preprocessing import StandardScaler

from sklearn.metrics import roc_auc_score
import sklearn.metrics as metrics

from sklearn.linear_model import LogisticRegression
from sklearn.ensemble import RandomForestClassifier
import lightgbm as lgb
import xgboost as xgb
from sklearn.svm import SVC
from sklearn.linear_model import SGDClassifier
from sklearn.calibration import CalibratedClassifierCV
from sklearn.neighbors import KNeighborsClassifier

# load processed train and validation data
train = pd.read_csv('../data/train_data_processed.csv')
valid = pd.read_csv('../data/valid_data_processed.csv')

# create X and y

X_train = train.drop(columns=['y'])
y_train = train['y']

X_valid = valid.drop(columns=['y'])
y_valid = valid['y']

print(X_train.shape)
print(y_train.shape)
print(X_valid.shape)
print(y_valid.shape)

## feature scaling

scaler = StandardScaler()

# fit on training data
scaler.fit(X_train)

# transform on train and valid
X_train_std = scaler.transform(X_train)
X_valid_std = scaler.transform(X_valid)

## baseline model - Logistic Regression

# starting with default parameters
lr = LogisticRegression(random_state=2021)

# fit the model
lr.fit(X_train_std, y_train)

# predictions
y_predict_train = lr.predict_proba(X_train_std)
y_predict_valid = lr.predict_proba(X_valid_std)

# AUC scores
train_score_lr = roc_auc_score(y_train, y_predict_train[:,1])
valid_score_lr = roc_auc_score(y_valid, y_predict_valid[:,1])
print("Training ROC-AUC score of baseline Logistic Regression model: ", train_score_lr)
print("Validation ROC-AUC score of baseline Logistic Regression model: ", valid_score_lr)

### observations:
Given that the [original paper](https://github.com/rachelkriggs/rocket/blob/main/docs/bank_telemarketing.pdf) regarding this problem stated that it achieved a best AUC score of 0.8, this results seems quite good for a simple baseline model with no hyperparameter tuning.

## Random Forest

# starting with default parameters
rf = RandomForestClassifier(random_state=2021)

# fit the model
rf.fit(X_train_std, y_train)

# predictions
y_predict_train = rf.predict_proba(X_train_std)
y_predict_valid = rf.predict_proba(X_valid_std)

# AUC scores
train_score_rf = roc_auc_score(y_train, y_predict_train[:,1])
valid_score_rf = roc_auc_score(y_valid, y_predict_valid[:,1])
print("Training ROC-AUC score of Random Forest model: ", train_score_rf)
print("Validation ROC-AUC score of Random Forest model: ", valid_score_rf)

### observations:
This model is clearly overfitting.

## LightGBM

# starting with default parameters
lgbm = lgb.LGBMClassifier(random_state = 2021, objective='binary')

# fit the model
lgbm.fit(X_train_std, y_train)

# predictions
y_predict_train = lgbm.predict_proba(X_train_std)
y_predict_valid = lgbm.predict_proba(X_valid_std)

# AUC scores
train_score_lgb = roc_auc_score(y_train, y_predict_train[:,1])
valid_score_lgb = roc_auc_score(y_valid, y_predict_valid[:,1])
print("Training ROC-AUC score of LightGBM model: ", train_score_lgb)
print("Validation ROC-AUC score of LightGBM model: ", valid_score_lgb)

## XGBoost

# starting with default parameters
xgbc = xgb.XGBClassifier(random_state = 2021, verbosity=0)

# fit the model
xgbc.fit(X_train_std, y_train)

# predictions
y_predict_train = xgbc.predict_proba(X_train_std)
y_predict_valid = xgbc.predict_proba(X_valid_std)

# AUC scores
train_score_xgb = roc_auc_score(y_train, y_predict_train[:,1])
valid_score_xgb = roc_auc_score(y_valid, y_predict_valid[:,1])
print("Training ROC-AUC score of XGBoost model: ", train_score_xgb)
print("Validation ROC-AUC score of XGBoost model: ", valid_score_xgb)

## Support Vector Machine

### RBF Kernel SVM (see the [scikit-learn documentation](https://scikit-learn.org/stable/modules/svm.html#classification))

# starting with default parameters
svcm = SVC(random_state = 2021, probability=True)

# fit the model
svcm.fit(X_train_std, y_train)

# predictions
y_predict_train = svcm.predict_proba(X_train_std)
y_predict_valid = svcm.predict_proba(X_valid_std)

# AUC scores
train_score_svc = roc_auc_score(y_train, y_predict_train[:,1])
valid_score_svc = roc_auc_score(y_valid, y_predict_valid[:,1])
print("Training ROC-AUC score of Support Vector Machine model: ", train_score_svc)
print("Validation ROC-AUC score of Support Vector Machine model: ", valid_score_svc)

### observations:
- The above model was quite slow to run, has a poor score compared to some of the others tested thus far, and is overfitting a fair bit.

### Linear SVM (see the [scikit-learn documentation](https://scikit-learn.org/stable/modules/svm.html#classification))

# starting with default parameters
lsvc = SGDClassifier(random_state = 2021)

# calibrate the model so we can get probability estimates
clsvc = CalibratedClassifierCV(lsvc)

# fit the model
clsvc.fit(X_train_std, y_train)

# predictions
y_predict_train = clsvc.predict_proba(X_train_std)
y_predict_valid = clsvc.predict_proba(X_valid_std)

# AUC scores
train_score_lsvc = roc_auc_score(y_train, y_predict_train[:,1])
valid_score_lsvc = roc_auc_score(y_valid, y_predict_valid[:,1])
print("Training ROC-AUC score of Linear SVM model: ", train_score_lsvc)
print("Validation ROC-AUC score of Linear SVM model: ", valid_score_lsvc)

## k-Nearest Neighbor

# starting with default parameters
knn = KNeighborsClassifier()
                          
# fit the model
knn.fit(X_train_std, y_train)

# predictions
y_predict_train = knn.predict_proba(X_train_std)
y_predict_valid = knn.predict_proba(X_valid_std)

# AUC scores
train_score_knn = roc_auc_score(y_train, y_predict_train[:,1])
valid_score_knn = roc_auc_score(y_valid, y_predict_valid[:,1])
print("Training ROC-AUC score of k-Nearest Neighbor model: ", train_score_knn)
print("Validation ROC-AUC score of k-Nearest Neighbor model: ", valid_score_knn)

## summary
of initial models tested with default parameters

summary = {
    'model':['Logistic Regression', 'Random Forest', 'LightGBM', 'XGBoost', 'RBF Kernel SVM', 'Linear SVM', 'k-Nearest Neighbor'],
    'Train ROC-AUC score':[train_score_lr, train_score_rf, train_score_lgb, train_score_xgb, train_score_svc, train_score_lsvc, train_score_knn],
    'Validation ROC-AUC score':[valid_score_lr, valid_score_rf, valid_score_lgb, valid_score_xgb, valid_score_svc, valid_score_lsvc, valid_score_knn]
}

pd.DataFrame(summary)

### observations:
From this intial run of models, candidates for further tuning are: 

- Logistic Regression
- LightGBM
- XGBoost
- Random Forest
