==> notebooks/04_tune.ipynb <==
Tune hyperparameters for the following models:

- Logistic Regression
- LightGBM
- XGBoost
- Random Forest

import pandas as pd
import numpy as np
import matplotlib.pyplot as plt 

from sklearn.preprocessing import StandardScaler

from sklearn.metrics import roc_auc_score
import sklearn.metrics as metrics
from sklearn.model_selection import  RandomizedSearchCV, GridSearchCV, cross_val_score

from sklearn.linear_model import LogisticRegression
from sklearn.ensemble import RandomForestClassifier
import lightgbm as lgb
import xgboost as xgb

# load processed train and validation data
train = pd.read_csv('../data/train_data_processed.csv')
valid = pd.read_csv('../data/valid_data_processed.csv')

# create X and y

X_train = train.drop(columns=['y'])
y_train = train['y']

X_valid = valid.drop(columns=['y'])
y_valid = valid['y']

print(X_train.shape)
print(y_train.shape)
print(X_valid.shape)
print(y_valid.shape)

##  feature scaling

scaler = StandardScaler()

# fit on training data
scaler.fit(X_train)

# transform on train and valid
X_train_std = scaler.transform(X_train)
X_valid_std = scaler.transform(X_valid)

## Hyperparameter Tuning

## 1. Logistic Regression

parameters = {
    'C': [0.001, 0.01, 0.1, 1.0, 10, 100],
    'solver': ['newton-cg', 'lbfgs', 'liblinear']
}

# define the model
lr = LogisticRegression(random_state=2021, max_iter=200, class_weight='balanced')
cv_lr = GridSearchCV(lr, parameters, cv=5)

# run the random search cv on the train set to find the best parameters
%time cv_lr.fit(X_train_std, y_train)

print('The best parameters are: \n', cv_lr.best_params_)

# fitting with best parameters from the above search

lr_model = LogisticRegression(
    random_state=2021, 
    max_iter=200,
    class_weight='balanced',
    C=0.1,  # best_params_ gives 100 but manual experimentation with this value was better
    solver='newton-cg') 

# fit the model
lr_model.fit(X_train_std, y_train)

# predictions  
y_predict_train = lr_model.predict_proba(X_train_std)
y_predict_valid = lr_model.predict_proba(X_valid_std)

# AUC scores
train_score_lr = roc_auc_score(y_train, y_predict_train[:,1])
valid_score_lr = roc_auc_score(y_valid, y_predict_valid[:,1])
print("Training ROC-AUC score of Logistic Regression model: ", train_score_lr)
print("Validation ROC-AUC score of Logistic Regression model: ", valid_score_lr)

## 2. LightGBM

# random search

parameters = {
    'num_leaves': [x for x in range(4,50, 3)],
    'max_depth':[x for x in range(-1,51, 3)],
    'learning_rate': [0.01, 0.015, 0.025, 0.05, 0.1, 0.3],
    'n_estimators': [x for x in range(100,1001, 100)],
}

# define the model
lgbm = lgb.LGBMClassifier(random_state=2021, objective='binary')
# cv_lgbm = GridSearchCV(lgbm, parameters, cv=5)
cv_lgbm = RandomizedSearchCV(lgbm, parameters, cv=5)

# run the random search cv on the train set to find the best parameters
%time cv_lgbm.fit(X_train_std, y_train)

print('The best parameters are: \n', cv_lgbm.best_params_)

# fitting with best parameters from the above search

lgb_model = lgb.LGBMClassifier(
    random_state=2021, 
    objective='binary',
    num_leaves=10,
    max_depth=11,
    learning_rate=0.01,
    n_estimators=900) 

# fit the model
lgb_model.fit(X_train_std, y_train)

# predictions
y_predict_train = lgb_model.predict_proba(X_train_std)
y_predict_valid = lgb_model.predict_proba(X_valid_std)

# AUC scores
train_score_lgb = roc_auc_score(y_train, y_predict_train[:,1])
valid_score_lgb = roc_auc_score(y_valid, y_predict_valid[:,1])
print("Training ROC-AUC score of LightGBM model: ", train_score_lgb)
print("Validation ROC-AUC score of LightGBM model: ", valid_score_lgb)

## 3. XGBoost

parameters = {
    'learning_rate': [0.01, 0.015, 0.025, 0.05, 0.1, 0.3],
    'n_estimators': [x for x in range(100,1001, 100)],
    'max_depth': [3, 4, 6, 8, 10],
    'min_child_weight': [1, 3, 5, 7],
    'subsample': [0.3, 0.4, 0.5, 0.6, 0.7, 0.8, 1],
    'colsample_bytree': [0.3, 0.4, 0.5, 0.6, 0.7, 0.8]
}

# define the model
xgbc = xgb.XGBClassifier(random_state = 2021, verbosity=0)
cv_xgb = RandomizedSearchCV(xgbc, parameters, cv=5)

# run the random search cv on the train set to find the best parameters
%time cv_xgb.fit(X_train_std, y_train)

print('The best parameters are: \n', cv_xgb.best_params_)

# fitting with best parameters from the above search

xgb_model = xgb.XGBClassifier(
    random_state=2021, 
    learning_rate=0.025,
    n_estimators=100,
    max_depth=10,
    min_child_weight=7,
    subsample=0.7,
    colsample_bytree=0.6, 
    verbosity=0
) 

# fit the model
xgb_model.fit(X_train_std, y_train)

# predictions
y_predict_train = xgb_model.predict_proba(X_train_std)
y_predict_valid = xgb_model.predict_proba(X_valid_std)

# AUC scores
train_score_xgb = roc_auc_score(y_train, y_predict_train[:,1])
valid_score_xgb = roc_auc_score(y_valid, y_predict_valid[:,1])
print("Training ROC-AUC score of XGBoost model: ", train_score_xgb)
print("Validation ROC-AUC score of XGBoost model: ", valid_score_xgb)

## 4. Random Forest

parameters = {
    'max_features': [x for x in range(15,(len(X_train.columns)+1), 50)],
    'n_estimators': [50, 100, 250, 500, 750, 1000], 
    'max_depth': [10, 25, 50, 75, 100] + [None]
}

# define the model
rf = RandomForestClassifier(random_state=2021)
cv_rf = RandomizedSearchCV(rf, parameters, cv=5)

# run the random search cv on the train set to find the best parameters
%time cv_rf.fit(X_train_std, y_train)

print('The best parameters are: \n', cv_rf.best_params_)

# fitting with best parameters from the above search

rf_model = RandomForestClassifier(
    random_state=2021, 
    n_estimators = 500, 
    max_features = 15,
    max_depth = 10
) 

# fit the model
rf_model.fit(X_train_std, y_train)

# predictions
y_predict_train = rf_model.predict_proba(X_train_std)
y_predict_valid = rf_model.predict_proba(X_valid_std)

# AUC scores
train_score_rf = roc_auc_score(y_train, y_predict_train[:,1])
valid_score_rf = roc_auc_score(y_valid, y_predict_valid[:,1])
print("Training ROC-AUC score of Random Forest model: ", train_score_rf)
print("Validation ROC-AUC score of Random Forest model: ", valid_score_rf)

## summary

summary = {
    'model':['Logistic Regression', 'LightGBM', 'XGBoost', 'Random Forest'],
    'Train ROC-AUC score':[train_score_lr, train_score_lgb, train_score_xgb, train_score_rf],
    'Validation ROC-AUC score':[valid_score_lr, valid_score_lgb, valid_score_xgb, valid_score_rf]
}

pd.DataFrame(summary)

### observations:

comparing these tuned models to the initial ones with default parameters in [this notebook](https://github.com/rachelkriggs/rocket/blob/main/notebooks/03_model.ipynb):

- The Logistic Regression model overfit a small amount more but overall performed about the same.
- The LightGBM model's validation score saw a small decrease; however, this model improved in overfitting less.
- The XGBoost model's validation score got better and this model overfit less.
- The Random Forest model's validation score increased and the model overfit less; however, this model is still overfitting the most in this set.
